Added Project 3 for Project_Showcase
The project is an implementation of a indexable Min Priority Queue that shows the lowest price car in the queue and the lowest mileage car in the queue. It also shows both the lowest mileage and price for a specified Make and Model. Data is pulled in from a text file and runs in O(log n) time.